cvhgfh

+++ Codigo/Bluepill/Core/Inc/sharp.h
@@ -11,7 +11,7 @@
 #include "main.h"
 
 uint32_t* readSensor(uint32_t *value);
-float lecturaSensor(int direccion, uint32_t *values);
+float lecturaSensor(int direccion, float *values);
 float lecSensor(int n, uint32_t *values);
 int lecturaCNY70(int n, uint32_t values[10]);
 int wallDetector(int n, int d);
